requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
itsdangerous>=2.2.0
bcrypt>=4.2.0
google-api-python-client>=2.147.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
resend>=2.0.0
pydantic-settings>=2.5.0
slowapi>=0.1.9
httpx>=0.27.0
icalendar>=5.0.0
pytest>=8.3.0
